Slightly de-hack CHStreamWriter using a holdSelf property

+++ Mac/Dolia/Dolia/CHDoliaListener.h
@@ -20,8 +20,8 @@
 @property (retain) NSNetService *service;
 @property NSObject<CHDoliaListenerDelegate> *delegate;
 @property (retain) NSMutableData *data;
-@property uint16_t length;
-@property uint16_t position;
+@property uint64_t length;
+@property NSUInteger position;
 - (void)stream:(NSStream *)theStream handleEvent:(NSStreamEvent)streamEvent;
 
 @end
